Face_Analysis/requirements.txt
tkinter
custom-tkinter
opencv-python==4.8.0
pillow==10.2.0
deepface==0.0.92
pyinstaller==6.9.0
